Sessions now show on refresh.
